+++ packages/wouter/lib/src/extensions/actions.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import '../delegate/delegate.dart';
 import '../models/models.dart';
 
@@ -9,11 +11,29 @@ extension RoutingActionsExtensions<T extends WouterDelegateState>
     }
 
     update((state) {
-      final prevState = policy.onPop(state);
+      var prevState;
 
-      if (predicate(prevState.fullPath)) {
-        return prevState;
-      }
+      do {
+        prevState = policy.onPop(state);
+      } while (predicate(prevState.fullPath));
+
+      return prevState;
     });
+  }
+
+  Future<T> popAndPush<T>(String path, [dynamic? result]) {
+    if (hasParent) {
+      return parent!.popAndPush(path, result);
+    }
+
+    final completer = Completer<T>();
+
+    update((state) {
+      final prevState = policy.onPop(state, result);
+
+      return policy.onPush(path, prevState, completer.complete);
+    });
+
+    return completer.future;
   }
 }